Flow refactoring
