+++ src/main/scala/Dispatcher.scala
@@ -11,7 +11,7 @@ object Dispatcher {
       f match {
         case "dns_post_lda" => dns_post_lda.run()
         case "dns_pre_lda" => dns_pre_lda.run()
-        case "flow_post_lda" =>flow_post_lda.run()
+        case "flow_post_lda" =>FlowPostLDA.run()
         case "flow_pre_lda" => FlowPreLDA.run()
       }
