+++ src/main/scala/NetFlowColumnIndex.scala
@@ -1,0 +1,25 @@
+package main.scala
+
+/**
+  * Created by rabarona on 6/2/2016.
+  */
+package main.scala {
+  object NetFlowColumnIndex extends Enumeration {
+    val HOUR = 4
+    val MINUTE = 5
+    val SECOND = 6
+    val SOURCEIP = 8
+    val DESTIP = 9
+    val SOURCEPORT = 10
+    val DESTPORT = 11
+    val IPKT = 16
+    val IBYT = 17
+    val NUMTIME = 27
+    val IBYTBIN = 28
+    val IPKTYBIN = 29
+    val TIMEBIN = 30
+    val SOURCEWORD = 33
+    val DESTWORD = 34
+  }
+}
+
